Add mock google-services.json

+++ app/src/main/java/hmatalonga/greenhub/managers/sampling/Inspector.java
@@ -615,8 +615,8 @@ public final class Inspector {
         long[] idleAndCpu2 = Cpu.readUsagePoint();
 
         // CPU status
-        double uptime = Cpu.getUptime();
-        double sleep = Cpu.getSleepTime();
+        long uptime = Cpu.getUptime();
+        long sleep = Cpu.getSleepTime();
 
         cpuStatus.cpuUsage = Cpu.getUsage(idleAndCpu1, idleAndCpu2);
         cpuStatus.upTime = uptime;
